requirements.txt
dill==0.3.1.1
requests==2.23.0
python-chess==0.30.1
py-cpuinfo==5.0.0
